requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.23.3
m3u8==3.5.0
ffmpeg-python==0.2.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.21.2
pyyaml==6.0.1